# src/main/resources/openapi.yaml
openapi: "3.0.3"
info:
  title: Employee Management API
  description: >
    REST API for managing employees. Provides CRUD operations for
    employee records including creation, retrieval, update, and deletion.
  version: "1.0.0"
  contact:
    name: EPAM Employee Management Team
servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Employee Management
    description: APIs for managing employees

paths:
  /api/employees:
    get:
      tags:
        - Employee Management
      summary: Get all employees
      description: Retrieves a list of all employees
      operationId: getAllEmployees
      responses:
        "200":
          description: Successfully retrieved list of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"

    post:
      tags:
        - Employee Management
      summary: Create a new employee
      description: Creates a new employee record
      operationId: createEmployee
      requestBody:
        description: Employee data to create
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeRequest"
      responses:
        "201":
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          description: Invalid input data

  /api/employees/{id}:
    get:
      tags:
        - Employee Management
      summary: Get employee by ID
      description: Retrieves a specific employee by their ID
      operationId: getEmployeeById
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the employee to retrieve
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Employee found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Employee not found
          content:
            application/json:
              schema:
                type: string
                example: "Employee not found with id: 1"

    put:
      tags:
        - Employee Management
      summary: Update an employee
      description: Updates an existing employee record
      operationId: updateEmployee
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the employee to update
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated employee data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeRequest"
      responses:
        "200":
          description: Employee updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          description: Invalid input data
        "404":
          description: Employee not found
          content:
            application/json:
              schema:
                type: string
                example: "Employee not found with id: 1"

    delete:
      tags:
        - Employee Management
      summary: Delete an employee
      description: Deletes an employee record by ID
      operationId: deleteEmployee
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the employee to delete
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Employee deleted successfully
        "404":
          description: Employee not found
          content:
            application/json:
              schema:
                type: string
                example: "Employee not found with id: 1"

components:
  schemas:
    Employee:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the employee
          example: 1
        firstName:
          type: string
          description: First name of the employee
          example: John
        lastName:
          type: string
          description: Last name of the employee
          example: Doe
        email:
          type: string
          format: email
          description: Email address of the employee
          example: john.doe@example.com
        department:
          type: string
          description: Department the employee belongs to
          example: Engineering
        position:
          type: string
          description: Job position of the employee
          example: Software Engineer
        salary:
          type: number
          format: double
          description: Salary of the employee
          example: 75000.00
      required:
        - id
        - firstName
        - lastName
        - email

    EmployeeRequest:
      type: object
      properties:
        firstName:
          type: string
          description: First name of the employee
          example: John
        lastName:
          type: string
          description: Last name of the employee
          example: Doe
        email:
          type: string
          format: email
          description: Email address of the employee
          example: john.doe@example.com
        department:
          type: string
          description: Department the employee belongs to
          example: Engineering
        position:
          type: string
          description: Job position of the employee
          example: Software Engineer
        salary:
          type: number
          format: double
          description: Salary of the employee
          example: 75000.00
      required:
        - firstName
        - lastName
        - email